v2.6.6
1. Add LTXV2 Frame Count node, calculate LTXV2-compatible frame count by rounding up to nearest multiple of 8 plus 1.
2. Add LTXV2 Cuts video gen workflow.
