+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "comfyui-jakeupgrade"
 description = "ComfyUI workflow customization by Jake."
-version = "2.6.5"
+version = "2.6.6"
 license = {file = "LICENSE"}
 dependencies = ["opencv-python", "piexif", "torch", "torchvision", "numpy", "pillow", "simpleeval", "PyYAML", "toml"]
